Added bootstrap, router, page header and split 2d and 3d into two pages
