+++ vis-tool/src/app/2d/visualisation/visualisation.2d.component.html
@@ -1,0 +1,18 @@
+<div class="col-xs-12" >
+  <div>
+    <div class="row">
+      <h2>2D Visualisation</h2>
+    </div>
+    <div class="row">
+      <h4>Select the JSON file to import</h4>
+    </div>
+    <input id="file-upload"
+           type="file"
+           accept=".txt"
+           (change)="fileUpload($event.target)">
+
+    <div class="row" style="margin-top: 100px">
+      <canvas #myCanvas style="border: 5px solid black;"></canvas>
+    </div>
+  </div>
+</div>
