+++ vis-tool/src/app/3d/visualisation/visualisation.3d.component.html
@@ -1,0 +1,3 @@
+<p>
+  Not yet implemented.
+</p>
