+++ vis-tool/src/app/3d/visualisation/visualisation.3d.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-visualisation',
+  templateUrl: './visualisation.3d.component.html',
+  styleUrls: ['./visualisation.3d.component.css']
+})
+export class Visualisation3dComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
