+++ vis-tool/src/app/app.component.html
@@ -1,15 +1,9 @@
-<div class="col-xs-12" >
-  <div >
-    <div class="row">
-      <h3>Select the .txt file to import</h3>
-    </div>
-    <input id="file-upload"
-           type="file"
-           accept=".txt"
-           (change)="fileUpload($event.target)">
+<app-header></app-header>
 
-    <div class="row" style="margin-top: 100px">
-      <canvas #myCanvas style="border: 5px solid black;"></canvas>
+<div class="container">
+  <div class="row">
+    <div class="col-md-12">
+      <router-outlet></router-outlet>
     </div>
   </div>
 </div>
