+++ vis-tool/src/app/header/header.component.html
@@ -1,0 +1,18 @@
+<nav class="navbar navbar-default">
+  <div class="container-fluid">
+    <div class="navbar-header">
+      <a routerLink="/" class="navbar-brand">Visualization Tool</a>
+    </div>
+
+    <div class="collapse navbar-collapse">
+      <ul class="nav navbar-nav navbar-right">
+          <li>
+            <a routerLink="/2d">2D</a>
+          </li>
+          <li>
+            <a routerLink="/3d">3D</a>
+          </li>
+      </ul>
+    </div>
+  </div>
+</nav>
